Changes in Logic Layer
I've created two folder:
Game and Game Event.
First one allows us to find a game by string  and to find which games are destined to be played during a GameEvent.
Second one  allows us to find a GameEvent by Creator Id and to Create it. It also has basic DTO classes.